Fix of input parameter size
The goal of this project was to practice with C / C++.  By setting up my dev environment, cmake, VSCode, Git, dependencies installation, I have successfully compiled this project and debugged a minor error.  processing the example data provided in the readme file will execute to completion after this fix is applied!

Hight and width were switched for different calls to Mat() when the source imC size had not changed.  When Height was entered as the first argument, the imC.cols size was different than the height which was caught in assertation checks.  This would have resulted in index out of range errors if the assertation did not catch this error.

This code fix will work on the example data set that is suggested in the readme.txt file for this project.  The code will not work for other data sets provided in this project.  Other assertation errors will occur.  This codebase is not release ready.